+++ SpringBootJFX/src/main/java/com/billing/dto/SupplierInvoiceDetails.java
@@ -22,7 +22,7 @@ private int stockNumber;
 	
 	private int noOfItems;
 	
-	private int totalQuanity;
+	private double totalQuanity;
 	
 	private double totalAmtWOTax;
 	
@@ -86,11 +86,11 @@ private int stockNumber;
 		this.noOfItems = noOfItems;
 	}
 
-	public int getTotalQuanity() {
+	public double getTotalQuanity() {
 		return totalQuanity;
 	}
 
-	public void setTotalQuanity(int totalQuanity) {
+	public void setTotalQuanity(double totalQuanity) {
 		this.totalQuanity = totalQuanity;
 	}
